refactor: move the logic of PostLengthFilter to a hook
