--- src/features/PostLengthFilter/types.ts
@@ -1,6 +1,0 @@
-export type LengthFilter = {
-  min?: number;
-  max?: number;
-};
-
-export type LengthFilterType = "min" | "max";
